[FEAT] ContestOutlineView UI 수정
