+++ Soongan/Soongan/Contest/VIew/GalleryView.swift
