+++ Soongan/Soongan/Contest/VIew/PhotoDetailView.swift
