+++ Soongan/Soongan/Contest/ViewModel/ContestOutlineViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  ContestOutlineViewModel.swift
+//  Soongan
+//
+//  Created by 김민 on 6/25/24.
+//
+
+import Foundation
+
+class ContestOutlineViewModel: ObservableObject {
+    
+    @Published var exhibitedCount = 3
+}
